+++ Interests/interestsRoutes.js
@@ -22,7 +22,12 @@ router.get('/', requireLogin, async (req, res) => {
 
 // Get the human readable name of an interest with :id
 router.get('/:id', requireLogin, async (req, res) => {
-    const { id } = req.params;
+    let { id } = req.params;
+
+    id = Number(id);
+
+    if(!id)
+        return res.status(400).json({ message: "An id number must be sent as a parameter." });
 
     try {
         const interest = await interestDB.getInterest(Number(id));
@@ -39,10 +44,13 @@ router.get('/:id', requireLogin, async (req, res) => {
 
 // Add a new interest to the database, interest name should be sent in the body. ID is auto generated.
 router.post('/', requireLogin, async (req, res) => {
-   const interest = req.body;
+   const { name } = req.body;
+
+   if(!name)
+       return res.status(400).json({ message: "An interest name must be sent in the body of the request." });
 
    try {
-       const result = await interestDB.addInterest(interest);
+       const result = await interestDB.addInterest({ name });
 
        if (result > 0)
            return res.status(200).json({ message: "Interest inserted" });
@@ -56,7 +64,12 @@ router.post('/', requireLogin, async (req, res) => {
 
 // Get all interests for the user with :user_id. Returns an array of interest ids
 router.get('/user/:user_id', requireLogin, async (req, res) => {
-    const { user_id } = req.params;
+    let { user_id } = req.params;
+
+    user_id = Number(user_id);
+
+    if(!user_id)
+        return res.status(400).json({ message: "A user id must be included as a parameter in the request." });
 
     try
     {
@@ -71,7 +84,12 @@ router.get('/user/:user_id', requireLogin, async (req, res) => {
 
 // Get all users who have :interest_id. Returns an array of user ids.
 router.get('/:interest_id/users', requireLogin, async (req, res) => {
-    const { interest_id } = req.params;
+    let { interest_id } = req.params;
+
+    interest_id = Number(interest_id);
+
+    if(!interest_id)
+        return res.status(400).json({ message: "An interest id must be included as a paramter in the request." });
 
     try
     {
@@ -86,13 +104,19 @@ router.get('/:interest_id/users', requireLogin, async (req, res) => {
 
 // Add an interest to a user. User id and interest id should be sent in the body of the request
 router.post('/user', requireLogin, async (req, res) => {
-    const { interest, user } = req.body;
+    let { interest, user } = req.body;
+
+    interest = Number(interest);
+    user = Number(user);
+
+    if(!interest || !user)
+        return res.status(400).json({ message: "A user id and an interest id must be included in the body of the request." });
 
     try
     {
         const result = await interestDB.addInterestToUser(user, interest);
 
-        if (result)
+        if (result > 0)
             return res.status(200).json({ message: "Interest added." });
         else
             return res.status(400).json({ message: "Unable to add interest." });
@@ -102,9 +126,15 @@ router.post('/user', requireLogin, async (req, res) => {
     }
 });
 
-// Remove an interest from a yser. User id and interest id should be sent in the body of the request
+// Remove an interest from a user. User id and interest id should be sent in the body of the request
 router.delete('/user', requireLogin, async (req, res) => {
-   const { user, interest } = req.body;
+   let { user, interest } = req.body;
+
+   user = Number(user);
+   interest = Number(interest);
+
+    if(!interest || !user)
+        return res.status(400).json({ message: "A user id and an interest id must be included in the body of the request." });
 
     try {
        const result = await interestDB.removeInterestFromUser(user, interest);
